3496 - Extract geo map setup logic to its own hook
